requirements.txt
appdirs==1.4.4
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.10
cssselect==1.1.0
docopt==0.6.2
fake-useragent==0.1.11
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.1
lxml==4.7.1
numpy==1.22.1
pandas==1.4.1
parse==1.19.0
pipreqs==0.4.11
psycopg==3.0.8
psycopg2-binary==2.9.3
pyee==8.2.2
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-html==0.10.0
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
tqdm==4.62.3
urllib3==1.26.8
w3lib==1.22.0
websockets==10.2
yarg==0.1.9
zipp==3.7.0
